% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecos.R
\name{ECOS_cleanup}
\alias{ECOS_cleanup}
\title{Clean up an ECOS workspace}
\usage{
ECOS_cleanup(workspace)
}
\arguments{
\item{workspace}{an external pointer of class \code{"ecos_workspace"}
as returned by \code{\link{ECOS_setup}}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Frees the C-level ECOS workspace. After this call the workspace
external pointer is invalidated and cannot be used for further
solves or updates. It is safe to call this more than once.
}
\seealso{
\code{\link{ECOS_setup}}, \code{\link{ECOS_solve}},
  \code{\link{ECOS_update}}
}
