% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECOS_exitcodes.R
\docType{data}
\name{ECOS_exitcodes}
\alias{ECOS_exitcodes}
\title{ECOS solver exit codes}
\description{
A two-column data frame consisting of the code and description for the
ECOS solver with ECOS symbolic code names as row names
}
\keyword{data}
