% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecos.R
\name{ECOS_solve}
\alias{ECOS_solve}
\title{Solve an ECOS workspace}
\usage{
ECOS_solve(workspace, control = NULL)
}
\arguments{
\item{workspace}{an external pointer of class \code{"ecos_workspace"}
as returned by \code{\link{ECOS_setup}}.}

\item{control}{optional named list of solver parameters (see
\code{\link{ecos.control}}). If \code{NULL} (the default), the
settings from \code{\link{ECOS_setup}} (or the most recent
\code{ECOS_solve} call with non-NULL control) are reused.}
}
\value{
the same result list as \code{\link{ECOS_csolve}}.
}
\description{
Calls \code{ECOS_solve} on an existing workspace created by
\code{\link{ECOS_setup}}. The workspace can be solved multiple
times, optionally with \code{\link{ECOS_update}} calls in between
to change numerical data.
}
\seealso{
\code{\link{ECOS_setup}}, \code{\link{ECOS_update}},
  \code{\link{ECOS_cleanup}}
}
