% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECOSolveR-package.R
\docType{package}
\name{ECOSolveR-package}
\alias{ECOSolveR}
\alias{ECOSolveR-package}
\title{ECOSolveR: Embedded Conic Solver in R}
\description{
\code{ECOSolveR} is a wrapper around the ecos library.
Please see the examples and documentation for the function \code{ECOS_csolve}.
}
\references{
\url{https://github.com/embotech/ecos}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bnaras.github.io/ECOSolveR/}
  \item Report bugs at \url{https://github.com/bnaras/ECOSolveR/issues}
}

}
\author{
\strong{Maintainer}: Balasubramanian Narasimhan \email{naras@stat.Stanford.EDU}

Authors:
\itemize{
  \item Anqi Fu \email{anqi.fu.1987@gmail.com}
}

Other contributors:
\itemize{
  \item Florian Schwendinger \email{FlorianSchwendinger@gmx.at} [contributor]
  \item Martin Maechler \email{maechler@stat.math.ethz.ch} [contributor]
}

}
\keyword{internal}
