% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECOSolveR.R
\docType{package}
\name{ECOSolveR}
\alias{ECOSolveR}
\title{ECOSolveR: Embedded Conic Solver in R}
\description{
\code{ECOSolveR} is a wrapper around the ecos library.
Please see the examples and documentation for the function \code{ECOS_csolve}.
}
\references{
\url{https://github.com/embotech/ecos}
}
