% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecos.R
\name{print.ecos_workspace}
\alias{print.ecos_workspace}
\title{Print method for ECOS workspace objects}
\usage{
\method{print}{ecos_workspace}(x, ...)
}
\arguments{
\item{x}{an \code{"ecos_workspace"} external pointer.}

\item{...}{ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for ECOS workspace objects
}
